Module re_arrow_store::store

Structs

ClusterCellCache

Used to cache auto-generated cluster cells ([0], [0, 1], [0, 1, 2], …) so that they can be properly deduplicated on insertion.

DataStore

A complete data store: covers all timelines, all entities, everything.

DataStoreConfig

DataTypeRegistry

Keeps track of datatype information for all component types that have been written to the store so far.

IndexedBucket

An IndexedBucket holds a chunk of rows from an IndexedTable (see DataStoreConfig::indexed_bucket_num_rows).

IndexedBucketInner

See IndexedBucket; this is a helper struct to simplify interior mutability.

IndexedTable

An IndexedTable is an ever-growing, arbitrary large re_log_types::DataTable that is optimized for time-based insertions and queries (which means a lot of bucketing).

MetadataRegistry

Keeps track of arbitrary per-row metadata.

PersistentIndexedTable

The timeless specialization of an IndexedTable.

StoreGeneration

Incremented on each edit

Type Definitions

InsertIdVec