Struct re_arrow_store::store::DataStore

pub struct DataStore {
    pub(crate) cluster_key: ComponentName,
    pub(crate) config: DataStoreConfig,
    pub(crate) messages: BTreeMap<MsgId, TimePoint>,
    pub(crate) cluster_comp_cache: IntMap<u32, RowIndex>,
    pub(crate) timeless_indices: IntMap<EntityPathHash, PersistentIndexTable>,
    pub(crate) timeless_components: IntMap<ComponentName, PersistentComponentTable>,
    pub(crate) indices: HashMap<(Timeline, EntityPathHash), IndexTable>,
    pub(crate) components: IntMap<ComponentName, ComponentTable>,
    pub(crate) insert_id: u64,
    pub(crate) query_id: AtomicU64,
    pub(crate) gc_id: u64,
}

A complete data store: covers all timelines, all entities, everything.

Debugging

DataStore provides a very thorough Display implementation that makes it manageable to know what’s going on internally. For even more information, you can set RERUN_DATA_STORE_DISPLAY_SCHEMAS=1 in your environment, which will result in additional schema information being printed out.

Additionally, if the polars feature is enabled, you can dump the entire datastore as a flat denormalized dataframe using Self::to_dataframe.

Fields
cluster_key: ComponentName

The cluster key specifies a column/component that is guaranteed to always be present for every single row of data within the store.

In addition to always being present, the payload of the cluster key..:

is always increasingly sorted,

is always dense (no validity bitmap),

and never contains duplicate entries.

This makes the cluster key a perfect candidate for joining query results together, and doing so as efficiently as possible.

See Self::insert_row for more information.

config: DataStoreConfig

The configuration of the data store (e.g. bucket sizes).

messages: BTreeMap<MsgId, TimePoint>

Maps MsgIds to some metadata (just timepoints at the moment).

BTreeMap because of garbage collection.

cluster_comp_cache: IntMap<u32, RowIndex>

Used to cache auto-generated cluster components, i.e. [0], [0, 1], [0, 1, 2], etc so that they can be properly deduplicated.

timeless_indices: IntMap<EntityPathHash, PersistentIndexTable>

Dedicated index tables for timeless data. Never garbage collected.

See also Self::indices.

timeless_components: IntMap<ComponentName, PersistentComponentTable>

Dedicated component tables for timeless data. Never garbage collected.

See also Self::components.

indices: HashMap<(Timeline, EntityPathHash), IndexTable>

Maps an entity to its index, for a specific timeline.

An index maps specific points in time to rows in component tables.

components: IntMap<ComponentName, ComponentTable>

Maps a component name to its associated table, for all timelines and all entities.

A component table holds all the values ever inserted for a given component.

insert_id: u64

Monotonically increasing ID for insertions.

query_id: AtomicU64

Monotonically increasing ID for queries.

gc_id: u64

Monotonically increasing ID for GCs.

Implementations

impl DataStore

pub fn new(cluster_key: ComponentName, config: DataStoreConfig) -> Self

See Self::cluster_key for more information about the cluster key.

pub fn insert_id_key() -> ComponentName

The column name used for storing insert requests’ IDs alongside the data.

The insert IDs are stored as-is directly into the index tables, this is not an indirection into an associated component table!

See DataStoreConfig::store_insert_ids.

pub fn cluster_key(&self) -> ComponentName

See Self::cluster_key for more information about the cluster key.

pub fn lookup_data_type(&self, component: &ComponentName) -> Option<&DataType>

Lookup the arrow DataType of a Component

impl DataStore

pub fn gc(
&mut self,
target: GarbageCollectionTarget,
primary_timeline: Timeline,
primary_component: ComponentName
) -> Vec<Box<dyn Array>>

Triggers a garbage collection according to the desired target, driven by the specified primary_component and primary_timeline. Returns all the raw data that was removed from the store for the given primary_component.

This only affects component tables, indices are left as-is, effectively behaving as tombstones.

The garbage collection is based on insertion order, which makes it both very efficient and very simple from an implementation standpoint. The tradeoff is that the given primary_timeline is expected to roughly follow insertion order, otherwise the behaviour is essentially undefined.

fn gc_drop_at_least_size_bytes(
&mut self,
primary_timeline: Timeline,
primary_component: ComponentName,
drop_at_least_size_bytes: f64
) -> Vec<Box<dyn Array>>

impl DataStore

pub fn all_components(
&self,
timeline: &Timeline,
ent_path: &EntityPath
) -> Option<Vec<ComponentName>>

Retrieve all the ComponentNames that have been written to for a given EntityPath on a specific Timeline.

Temporal semantics

In addition to the temporal results, this also includes all ComponentNames present in the timeless indices for this entity.

pub fn latest_at<const N: usize>(
&self,
query: &LatestAtQuery,
ent_path: &EntityPath,
primary: ComponentName,
components: &[ComponentName; N]
) -> Option<[Option<RowIndex>; N]>

Queries the datastore for the internal row indices of the specified components, as seen from the point of view of the so-called primary component.

Returns an array of row indices on success, or None otherwise. Success is defined by one thing and thing only: whether a row index could be found for the primary component. The presence or absence of secondary components has no effect on the success criteria.

On success, the returned array is filled with the internal row index of each and every component in components, or None if said component is not available in that row.

To actually retrieve the data associated with these indices, see Self::get.

Temporal semantics

Temporal indices take precedence, then timeless indices are queried to fill the holes left by missing temporal data.

Example

The following example demonstrate how to fetch the latest row indices for a given component and the associated cluster key, then get the corresponding data using these row indices, and finally turn everything into a nice-to-work-with polars’s dataframe.

pub fn latest_component(
    store: &DataStore,
    query: &LatestAtQuery,
    ent_path: &EntityPath,
    primary: ComponentName,
) -> anyhow::Result<DataFrame> {
    let cluster_key = store.cluster_key();

    let components = &[cluster_key, primary];
    let row_indices = store
        .latest_at(query, ent_path, primary, components)
        .unwrap_or([None; 2]);
    let results = store.get(components, &row_indices);

    let series: Result<Vec<_>, _> = components
        .iter()
        .zip(results)
        .filter_map(|(component, col)| col.map(|col| (component, col)))
        .map(|(&component, col)| Series::try_from((component.as_str(), col)))
        .collect();

    DataFrame::new(series?).map_err(Into::into)
}

Thanks to the cluster key, one is free to repeat this process as many times as they wish, then reduce the resulting dataframes down to one by joining them as they see fit. This is what our latest_components polars helper does.

For more information about working with dataframes, see the polars feature.

pub fn range<'a, const N: usize>(
&'a self,
query: &RangeQuery,
ent_path: &EntityPath,
components: [ComponentName; N]
) -> impl Iterator<Item = (Option<TimeInt>, IndexRowNr, [Option<RowIndex>; N])> + 'a

Iterates the datastore in order to return the internal row indices of the the specified components, as seen from the point of view of the so-called primary component, for the given time range.

For each and every relevant row that is found, the returned iterator will yield an array that is filled with the internal row index of each and every component in components, or None if said component is not available in that row. A row is considered iff it contains data for the primary component.

This method cannot fail! If there’s no data to return (whether that’s due to a missing primary index, missing secondary components, an empty point-of-view…), then an empty iterator is returned.

To actually retrieve the data associated with these indices, see Self::get.

⚠ Contrary to latest-at queries, range queries can and will yield multiple rows for a single timestamp if that timestamp happens to hold multiple entries for the primary component. On the contrary, they won’t yield any rows that don’t contain an actual value for the primary component, even if said rows do contain a value for one the secondaries!

Temporal semantics

Yields the contents of the temporal indices. Iff the query’s time range starts at TimeInt::MIN, this will yield the contents of the timeless indices before anything else.

When yielding timeless entries, the associated time will be None.

Example

The following example demonstrate how to range over the row indices of a given component and its associated cluster key, then get the corresponding data using these row indices, and finally turn everything into a nice-to-work-with iterator of polars’s dataframe. Additionally, it yields the latest-at state of the component a the start of the time range, if available.

pub fn range_component<'a>(
    store: &'a DataStore,
    query: &'a RangeQuery,
    ent_path: &'a EntityPath,
    primary: ComponentName,
) -> impl Iterator<Item = anyhow::Result<(Option<TimeInt>, DataFrame)>> + 'a {
    let cluster_key = store.cluster_key();

    let components = [cluster_key, primary];

    // Fetch the latest-at data just before the start of the time range.
    let latest_time = query.range.min.as_i64().saturating_sub(1).into();
    let df_latest = {
        let query = LatestAtQuery::new(query.timeline, latest_time);
        let row_indices = store
            .latest_at(&query, ent_path, primary, &components)
            .unwrap_or([None; 2]);
        let results = store.get(&components, &row_indices);
        dataframe_from_results(&components, results)
    };

    // Send the latest-at state before anything else..
    std::iter::once(df_latest.map(|df| (Some(latest_time), df)))
        // ..but only if it's not an empty dataframe.
        .filter(|df| df.as_ref().map_or(true, |(_, df)| !df.is_empty()))
        .chain(store.range(query, ent_path, components).map(
            move |(time, _, row_indices)| {
                let results = store.get(&components, &row_indices);
                dataframe_from_results(&components, results).map(|df| (time, df))
            },
        ))
}

Thanks to the cluster key, one is free to repeat this process as many times as they wish, then join the resulting streams to yield a full-fledged dataframe for every update of the primary component. This is what our range_components polars helper does.

For more information about working with dataframes, see the polars feature.

pub fn get<const N: usize>(
&self,
components: &[ComponentName; N],
row_indices: &[Option<RowIndex>; N]
) -> [Option<Box<dyn Array>>; N]

Retrieves the data associated with a list of components at the specified indices.

If the associated data is found, it will be written into the returned array at the appropriate index, or None otherwise.

row_indices takes a list of options so that one can easily re-use the results obtained from Self::latest_at & Self::range.

pub fn get_msg_metadata(&self, msg_id: &MsgId) -> Option<&TimePoint>

pub fn sort_indices_if_needed(&mut self)

Sort all unsorted indices in the store.

pub fn iter_indices(
&self
) -> impl ExactSizeIterator<Item = ((Timeline, EntityPath), &IndexTable)>

Returns a read-only iterator over the raw index tables.

Do not use this to try and test the internal state of the datastore.

impl DataStore

pub fn sanity_check(&self) -> Result<()>

Runs the sanity check suite for the entire datastore.

Returns an error if anything looks wrong.

pub fn oldest_time_per_timeline(&self) -> BTreeMap<Timeline, TimeInt>

The oldest time for which we have any data.

Ignores timeless data.

Useful to call after a gc.

impl DataStore

pub fn total_timeless_index_rows(&self) -> u64

Returns the number of timeless index rows stored across this entire store, i.e. the sum of the number of rows across all of its timeless index tables.

pub fn total_timeless_index_size_bytes(&self) -> u64

Returns the size of the timeless index data stored across this entire store, i.e. the sum of the size of the data stored across all of its timeless index tables, in bytes.

pub fn total_timeless_component_rows(&self) -> u64

Returns the number of timeless component rows stored across this entire store, i.e. the sum of the number of rows across all of its timeless component tables.

pub fn total_timeless_component_size_bytes(&self) -> u64

Returns the size of the timeless component data stored across this entire store, i.e. the sum of the size of the data stored across all of its timeless component tables, in bytes.

pub fn total_temporal_index_rows(&self) -> u64

Returns the number of temporal index rows stored across this entire store, i.e. the sum of the number of rows across all of its temporal index tables.

pub fn total_temporal_index_size_bytes(&self) -> u64

Returns the size of the temporal index data stored across this entire store, i.e. the sum of the size of the data stored across all of its temporal index tables, in bytes.

pub fn total_temporal_index_buckets(&self) -> u64

Returns the number of temporal index buckets stored across this entire store.

pub fn total_temporal_component_rows(&self) -> u64

Returns the number of temporal component rows stored across this entire store, i.e. the sum of the number of rows across all of its temporal component tables.

pub fn total_temporal_component_size_bytes(&self) -> u64

Returns the size of the temporal component data stored across this entire store, i.e. the sum of the size of the data stored across all of its temporal component tables, in bytes.

pub fn total_temporal_component_buckets(&self) -> u64

Returns the number of temporal component buckets stored across this entire store.

impl DataStore

pub fn insert_table(&mut self, table: &DataTable) -> WriteResult<()>

Inserts a [DataTable]’s worth of components into the datastore.

This iteratively inserts all rows from the table on a row-by-row basis. The entire method fails if any row fails.

Both the write and read paths transparently benefit from the contiguous memory of the table’s columns: the bigger the tables, the bigger the benefits!

See Self::insert_row.

pub fn insert_row(&mut self, row: &DataRow) -> WriteResult<()>

Inserts a [DataRow]’s worth of components into the datastore.

If the bundle doesn’t carry a payload for the cluster key, one will be auto-generated based on the length of the components in the payload, in the form of an array of monotonically increasing u64s going from 0 to N-1.

fn insert_timeless_row_helper(
&mut self,
cluster_cell_pos: Option<usize>,
cells: &[DataCell],
row_indices: &mut IntMap<ComponentName, RowIndex>
) -> WriteResult<()>

fn insert_row_helper(
&mut self,
time_point: &TimePoint,
cluster_cell_pos: Option<usize>,
cells: &[DataCell],
row_indices: &mut IntMap<ComponentName, RowIndex>
) -> WriteResult<()>

fn get_or_create_cluster_component(
&mut self,
cluster_cell_pos: Option<usize>,
cells: &[DataCell],
time_point: &TimePoint
) -> WriteResult<RowIndex>

Tries to find the cluster component for the current row, or creates it if the caller hasn’t specified any.

When creating an auto-generated cluster component of a specific length for the first time, this will keep track of its assigned row index and re-use it later on as a mean of deduplication.

pub fn clear_msg_metadata(&mut self, drop_msg_ids: &HashSet<MsgId>)

impl DataStore

pub fn to_dataframe(&self) -> DataFrame

Dumps the entire datastore as a flat, denormalized dataframe.

This cannot fail: it always tries to yield as much valuable information as it can, even in the face of errors.