Trait re_arrow_store::ArrayExt

pub trait ArrayExt: Array {
    fn get_child_length(&self, child_nr: usize) -> usize;
    fn clean_for_polars(&self) -> Box<dyn Array>;
}

Required Methods

fn get_child_length(&self, child_nr: usize) -> usize

Returns the length of the child array at the given index.

Panics if self is not a ListArray<i32>.

Panics if child_nr is out of bounds.

fn clean_for_polars(&self) -> Box<dyn Array>

Create a new Array which avoids problematic types for polars.

This does the following conversion:

FixedSizeList -> List

Union -> Struct

Nested types are expanded and cleaned recursively