Struct re_arrow_store::store::DataStore

pub struct DataStore {
    pub(crate) cluster_key: ComponentName,
    pub(crate) config: DataStoreConfig,
    pub(crate) type_registry: DataTypeRegistry,
    pub(crate) metadata_registry: MetadataRegistry<TimePoint>,
    pub(crate) cluster_cell_cache: ClusterCellCache,
    pub(crate) tables: HashMap<(Timeline, EntityPathHash), IndexedTable>,
    pub(crate) timeless_tables: IntMap<EntityPathHash, PersistentIndexedTable>,
    pub(crate) insert_id: u64,
    pub(crate) query_id: AtomicU64,
    pub(crate) gc_id: u64,
}

A complete data store: covers all timelines, all entities, everything.

Debugging

DataStore provides a very thorough Display implementation that makes it manageable to know what’s going on internally. For even more information, you can set RERUN_DATA_STORE_DISPLAY_SCHEMAS=1 in your environment, which will result in additional schema information being printed out.

Additionally, if the polars feature is enabled, you can dump the entire datastore as a flat denormalized dataframe using Self::to_dataframe.

Fields
cluster_key: ComponentName

The cluster key specifies a column/component that is guaranteed to always be present for every single row of data within the store.

In addition to always being present, the payload of the cluster key..:

is always increasingly sorted,

is always dense (no validity bitmap),

and never contains duplicate entries.

This makes the cluster key a perfect candidate for joining query results together, and doing so as efficiently as possible.

See Self::insert_row for more information.

config: DataStoreConfig

The configuration of the data store (e.g. bucket sizes).

type_registry: DataTypeRegistry

Keeps track of datatype information for all component types that have been written to the store so far.

See also Self::lookup_datatype.

metadata_registry: MetadataRegistry<TimePoint>

Keeps track of arbitrary per-row metadata.

Only used to map RowIds to their original TimePoints at the moment.

cluster_cell_cache: ClusterCellCache

Used to cache auto-generated cluster cells ([0], [0, 1], [0, 1, 2], …) so that they can be properly deduplicated on insertion.

tables: HashMap<(Timeline, EntityPathHash), IndexedTable>

All temporal IndexedTables for all entities on all timelines.

See also Self::timeless_tables.

timeless_tables: IntMap<EntityPathHash, PersistentIndexedTable>

All timeless indexed tables for all entities. Never garbage collected.

See also Self::tables.

insert_id: u64

Monotonically increasing ID for insertions.

query_id: AtomicU64

Monotonically increasing ID for queries.

gc_id: u64

Monotonically increasing ID for GCs.

Implementations

impl DataStore

pub fn new(cluster_key: ComponentName, config: DataStoreConfig) -> Self

See Self::cluster_key for more information about the cluster key.

pub fn insert_id_key() -> ComponentName

The column name used for storing insert requests’ IDs alongside the data when manipulating dataframes.

See DataStoreConfig::store_insert_ids.

pub fn generation(&self) -> StoreGeneration

Return the current StoreGeneration. This can be used to determine whether the database has been modified since the last time it was queried.

pub fn cluster_key(&self) -> ComponentName

See Self::cluster_key for more information about the cluster key.

pub fn config(&self) -> &DataStoreConfig

See DataStoreConfig for more information about configuration.

pub fn lookup_datatype(&self, component: &ComponentName) -> Option<&DataType>

Lookup the arrow [DataType] of a re_types::Component in the internal DataTypeRegistry.

pub fn oldest_time_per_timeline(&self) -> BTreeMap<Timeline, TimeInt>

The oldest time for which we have any data.

Ignores timeless data.

Useful to call after a gc.

pub fn iter_indices( &self ) -> impl ExactSizeIterator<Item = ((Timeline, EntityPath), &IndexedTable)>

Returns a read-only iterator over the raw indexed tables.

Do not use this to try and assert the internal state of the datastore.

impl DataStore

pub fn to_data_table(&self) -> DataReadResult<DataTable>

Serializes the entire datastore into one big sorted DataTable.

Individual re_log_types::DataRows that were split apart due to bucketing are merged back together.

Beware: this is extremely costly, don’t use this in hot paths.

pub fn to_data_tables( &self, time_filter: Option<(Timeline, TimeRange)> ) -> impl Iterator<Item = DataTable> + '_

Serializes the entire datastore into an iterator of DataTables, where each table corresponds 1-to-1 to an internal bucket.

fn dump_timeless_tables(&self) -> impl Iterator<Item = DataTable> + '_

fn dump_temporal_tables(&self) -> impl Iterator<Item = DataTable> + '_

fn dump_temporal_tables_filtered( &self, (timeline_filter, time_filter): (Timeline, TimeRange) ) -> impl Iterator<Item = DataTable> + '_

impl DataStore

pub fn gc( &mut self, options: GarbageCollectionOptions ) -> (Deleted, DataStoreStats)

Triggers a garbage collection according to the desired target.

Garbage collection’s performance is bounded by the number of buckets in each table (for each RowId, we have to find the corresponding bucket, which is roughly O(log(n))) as well as the number of rows in each of those buckets (for each RowId, we have to sort the corresponding bucket (roughly O(n*log(n))) and then find the corresponding row (roughly O(log(n)). The size of the data itself has no impact on performance.

Returns the list of RowIds that were purged from the store.

Semantics

Garbage collection works on a row-level basis and is driven by RowId order, i.e. the order defined by the clients’ wall-clocks, allowing it to drop data across the different timelines in a fair, deterministic manner. Similarly, out-of-order data is supported out of the box.

The garbage collector doesn’t deallocate data in and of itself: all it does is drop the store’s internal references to that data (the DataCells), which will be deallocated once their reference count reaches 0.

Limitations

The garbage collector has limited support for latest-at semantics. The configuration option: GarbageCollectionOptions::protect_latest will protect the N latest values of each component on each timeline. The only practical guarantee this gives is that a latest-at query with a value of max-int will be be unchanged. However, latest-at queries from other arbitrary points in time may provide different results pre- and post- GC.

NOTE: This configuration option is not yet enabled for the Rerun viewer GC pass.

See https://github.com/rerun-io/rerun/issues/1803.

fn gc_drop_at_least_num_bytes( &mut self, num_bytes_to_drop: f64, include_timeless: bool, protected_rows: &HashSet<RowId> ) -> Deleted

Tries to drop at least num_bytes_to_drop bytes of data from the store.

Returns the list of RowIds that were purged from the store.

fn find_all_protected_rows(&mut self, target_count: usize) -> HashSet<RowId>

For each EntityPath, Timeline, Component find the N latest RowIds.

These are the rows that must be protected so as not to impact a latest-at query. Note that latest for Timeless is currently based on insertion-order rather than tuid. See: #1807

fn purge_empty_tables(&mut self) -> Vec<RowId>

Remove any tables which contain only components which are empty.

impl DataStore

pub fn query_latest_component<C: Component>( &self, entity_path: &EntityPath, query: &LatestAtQuery ) -> Option<VersionedComponent<C>>

Get the latest value for a given re_types::Component and the associated RowId.

This assumes that the row we get from the store only contains a single instance for this component; it will log a warning otherwise.

This should only be used for “mono-components” such as Transform and Tensor.

This is a best-effort helper, it will merely log errors on failure.

pub fn query_latest_component_at_closest_ancestor<C: Component>( &self, entity_path: &EntityPath, query: &LatestAtQuery ) -> Option<(EntityPath, VersionedComponent<C>)>

Call Self::query_latest_component at the given path, walking up the hierarchy until an instance is found.

pub fn query_timeless_component<C: Component>( &self, entity_path: &EntityPath ) -> Option<VersionedComponent<C>>

Get the latest value for a given re_types::Component and the associated RowId, assuming it is timeless.

This assumes that the row we get from the store only contains a single instance for this component; it will log a warning otherwise.

This should only be used for “mono-components” such as Transform and Tensor.

This is a best-effort helper, it will merely log errors on failure.

impl DataStore

pub fn insert_component<'a, C>( &mut self, entity_path: &EntityPath, timepoint: &TimePoint, component: C )where C: Component + Clone + 'a, Cow<'a, C>: From<C>,

Stores a single value for a given re_types::Component.

This is a best-effort helper, it will merely log errors on failure.

pub fn insert_empty_component( &mut self, entity_path: &EntityPath, timepoint: &TimePoint, component: ComponentName )

Stores a single empty value for a given re_log_types::ComponentName.

This is a best-effort helper, it will merely log errors on failure.

impl DataStore

pub fn all_components( &self, timeline: &Timeline, ent_path: &EntityPath ) -> Option<Vec<ComponentName>>

Retrieve all the ComponentNames that have been written to for a given EntityPath on a specific Timeline.

Temporal semantics

In addition to the temporal results, this also includes all ComponentNames present in the timeless tables for this entity.

pub fn entity_has_component( &self, timeline: &Timeline, ent_path: &EntityPath, component: &ComponentName ) -> bool

Check whether a given entity has a specific ComponentName on the specified timeline.

Temporal semantics

In addition to the temporal results, this also checks whether the ComponentName is present in timeless table.

pub fn latest_at<const N: usize>( &self, query: &LatestAtQuery, ent_path: &EntityPath, primary: ComponentName, components: &[ComponentName; N] ) -> Option<(RowId, [Option<DataCell>; N])>

Queries the datastore for the cells of the specified components, as seen from the point of view of the so-called primary component.

Returns an array of DataCells on success, or None otherwise. Success is defined by one thing and thing only: whether a cell could be found for the primary component. The presence or absence of secondary components has no effect on the success criteria.

Temporal semantics

Temporal indices take precedence, then timeless tables are queried to fill the holes left by missing temporal data.

Example

The following example demonstrate how to fetch the latest cells for a given component and its associated cluster key, and wrap the result into a nice-to-work-with polars’s dataframe.

pub fn latest_component(
    store: &DataStore,
    query: &LatestAtQuery,
    ent_path: &EntityPath,
    primary: ComponentName,
) -> anyhow::Result<DataFrame> {
    let cluster_key = store.cluster_key();

    let components = &[cluster_key, primary];
    let (_, cells) = store
        .latest_at(&query, ent_path, primary, components)
        .unwrap_or((RowId::ZERO, [(); 2].map(|_| None)));

    let series: Result<Vec<_>, _> = cells
        .iter()
        .flatten()
        .map(|cell| {
            Series::try_from((
                cell.component_name().as_str(),
                cell.to_arrow(),
            ))
        })
        .collect();

    DataFrame::new(series?).map_err(Into::into)
}

Thanks to the cluster key, one is free to repeat this process as many times as they wish, then reduce the resulting dataframes down to one by joining them as they see fit. This is what our latest_components polars helper does.

For more information about working with dataframes, see the polars feature.

pub fn range<'a, const N: usize>( &'a self, query: &RangeQuery, ent_path: &EntityPath, components: [ComponentName; N] ) -> impl Iterator<Item = (Option<TimeInt>, RowId, [Option<DataCell>; N])> + 'a

Iterates the datastore in order to return the cells of the specified components, as seen from the point of view of the so-called primary component, for the given time range.

For each and every relevant row that is found, the returned iterator will yield an array that is filled with the cells of each and every component in components, or None if said component is not available in that row. A row is considered iff it contains data for the primary component.

This method cannot fail! If there’s no data to return, an empty iterator is returned.

⚠ Contrary to latest-at queries, range queries can and will yield multiple rows for a single timestamp if that timestamp happens to hold multiple entries for the primary component. On the contrary, they won’t yield any rows that don’t contain an actual value for the primary component, even if said rows do contain a value for one the secondaries!

Temporal semantics

Yields the contents of the temporal indices. Iff the query’s time range starts at TimeInt::MIN, this will yield the contents of the timeless tables before anything else.

When yielding timeless entries, the associated time will be None.

Example

The following example demonstrate how to range over the cells of a given component and its associated cluster key, and turn the results into a nice-to-work-with iterator of polars’s dataframe. Additionally, it yields the latest-at state of the component at the start of the time range, if available.

pub fn range_component<'a>(
    store: &'a DataStore,
    query: &'a RangeQuery,
    ent_path: &'a EntityPath,
    primary: ComponentName,
) -> impl Iterator<Item = anyhow::Result<(Option<TimeInt>, DataFrame)>> + 'a {
    let cluster_key = store.cluster_key();

    let components = [cluster_key, primary];

    // Fetch the latest-at data just before the start of the time range.
    let latest_time = query.range.min.as_i64().saturating_sub(1).into();
    let df_latest = {
        let query = LatestAtQuery::new(query.timeline, latest_time);
        let (_, cells) = store
            .latest_at(&query, ent_path, primary, &components)
            .unwrap_or((RowId::ZERO, [(); 2].map(|_| None)));
        dataframe_from_cells(cells)
    };

    // Send the latest-at state before anything else..
    std::iter::once(df_latest.map(|df| (Some(latest_time), df)))
        // ..but only if it's not an empty dataframe.
        .filter(|df| df.as_ref().map_or(true, |(_, df)| !df.is_empty()))
        .chain(store.range(query, ent_path, components).map(
            move |(time, _, cells)| dataframe_from_cells(cells).map(|df| (time, df))
        ))
}

Thanks to the cluster key, one is free to repeat this process as many times as they wish, then join the resulting streams to yield a full-fledged dataframe for every update of the primary component. This is what our range_components polars helper does.

For more information about working with dataframes, see the polars feature.

pub fn get_msg_metadata(&self, row_id: &RowId) -> Option<&TimePoint>

pub fn sort_indices_if_needed(&mut self)

Sort all unsorted indices in the store.

impl DataStore

pub fn sanity_check(&self) -> Result<(), SanityError>

Runs the sanity check suite for the entire datastore.

Returns an error if anything looks wrong.

impl DataStore

pub fn num_timeless_rows(&self) -> u64

Returns the number of timeless index rows stored across this entire store, i.e. the sum of the number of rows across all of its timeless indexed tables.

pub fn timeless_size_bytes(&self) -> u64

Returns the size of the timeless index data stored across this entire store, i.e. the sum of the size of the data stored across all of its timeless indexed tables, in bytes.

pub fn num_temporal_rows(&self) -> u64

Returns the number of temporal index rows stored across this entire store, i.e. the sum of the number of rows across all of its temporal indexed tables.

pub fn temporal_size_bytes(&self) -> u64

Returns the size of the temporal index data stored across this entire store, i.e. the sum of the size of the data stored across all of its temporal indexed tables, in bytes.

pub fn num_temporal_buckets(&self) -> u64

Returns the number of temporal indexed buckets stored across this entire store.

pub fn entity_stats( &self, timeline: Timeline, entity_path_hash: EntityPathHash ) -> EntityStats

Stats for a specific entity path on a specific timeline

impl DataStore

pub fn insert_table(&mut self, table: &DataTable) -> WriteResult<()>

Inserts a DataTable’s worth of components into the datastore.

This iteratively inserts all rows from the table on a row-by-row basis. The entire method fails if any row fails.

Both the write and read paths transparently benefit from the contiguous memory of the table’s columns: the bigger the tables, the bigger the benefits!

See Self::insert_row.

pub fn insert_row(&mut self, row: &DataRow) -> WriteResult<()>

Inserts a DataRow’s worth of components into the datastore.

If the bundle doesn’t carry a payload for the cluster key, one will be auto-generated based on the length of the components in the payload, in the form of an array of monotonically increasing u64s going from 0 to N-1.

pub fn wipe_timeless_data(&mut self)

Wipes all timeless data.

Mostly useful for testing/debugging purposes.

fn generate_cluster_cell(&mut self, num_instances: u32) -> DataCell

Auto-generates an appropriate cluster cell for the specified number of instances and transparently handles caching.

impl DataStore

pub fn to_dataframe(&self) -> DataFrame

Dumps the entire datastore as a flat, denormalized dataframe.

This cannot fail: it always tries to yield as much valuable information as it can, even in the face of errors.